refactor: use Plugin object
