+++ composer.json
@@ -28,9 +28,7 @@
         "cakephp/twig-view": "^1.0.1"
     },
     "require-dev": {
-        "cakephp/bake": "^2.0.3",
         "cakephp/cakephp-codesniffer": "~4.0.0",
-        "cakephp/debug_kit": "^4.0",
         "phpunit/phpunit": "^8.0"
     },
     "autoload": {
