--- src/BaseApplication.php
@@ -1,110 +1,0 @@
-<?php
-declare(strict_types=1);
-
-/**
- * BEdita, API-first content management framework
- * Copyright 2018 ChannelWeb Srl, Chialab Srl
- *
- * This file is part of BEdita: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * See LICENSE.LGPL or <http://gnu.org/licenses/lgpl-3.0.html> for more details.
- */
-namespace BEdita\WebTools;
-
-use BEdita\WebTools\Command\CacheClearallCommand;
-use Cake\Console\CommandCollection;
-use Cake\Core\Configure;
-use Cake\Core\Exception\MissingPluginException;
-use Cake\Error\Middleware\ErrorHandlerMiddleware;
-use Cake\Http\BaseApplication as CakeBaseApplication;
-use Cake\Http\MiddlewareQueue;
-use Cake\Routing\Middleware\AssetMiddleware;
-use Cake\Routing\Middleware\RoutingMiddleware;
-
-/**
- * Application setup base class.
- *
- * This defines the bootstrapping logic and middleware layers you
- * want to use in your application.
- */
-class BaseApplication extends CakeBaseApplication
-{
-    /**
-     * Use `cache clear_all` from BEdita\WebTools\Command\CacheClearallCommand
-     *
-     * @param \Cake\Console\CommandCollection $commands Console commands.
-     * @return \Cake\Console\CommandCollection
-     */
-    public function console(CommandCollection $commands): CommandCollection
-    {
-        parent::console($commands);
-        $commands->remove('cache clear_all');
-        $commands->add('cache clear_all', CacheClearallCommand::class);
-
-        return $commands;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function bootstrap(): void
-    {
-        // Call parent to load bootstrap from files.
-        parent::bootstrap();
-
-        if (PHP_SAPI === 'cli') {
-            $this->bootstrapCli();
-        }
-
-        /*
-         * Only try to load DebugKit in development mode
-         * Debug Kit should not be installed on a production system
-         */
-        if (Configure::read('debug')) {
-            try {
-                $this->addPlugin('DebugKit');
-            } catch (MissingPluginException $e) {
-                // Do not halt if the plugin is missing
-            }
-        }
-        // Load more plugins here
-        $this->addPlugin('Cake/TwigView');
-    }
-
-    /**
-     * Setup the middleware queue your application will use.
-     *
-     * @param \Cake\Http\MiddlewareQueue $middlewareQueue The middleware queue to setup.
-     * @return \Cake\Http\MiddlewareQueue The updated middleware queue.
-     */
-    public function middleware(MiddlewareQueue $middlewareQueue): MiddlewareQueue
-    {
-        $middlewareQueue
-            // Catch any exceptions in the lower layers,
-            // and make an error page/response
-            ->add(ErrorHandlerMiddleware::class)
-
-            // Handle plugin/theme assets like CakePHP normally does.
-            ->add(AssetMiddleware::class)
-
-            // Add routing middleware.
-            ->add(new RoutingMiddleware($this));
-
-        return $middlewareQueue;
-    }
-
-    /**
-     * @return void
-     */
-    protected function bootstrapCli(): void
-    {
-        try {
-            $this->addPlugin('Bake');
-        } catch (MissingPluginException $e) {
-            // Do not halt if the plugin is missing
-        }
-    }
-}
